Switch to use email as username
The app was getting unnecessarily complex using a username because then things
like password reset wouldn't work unless we had an email. Also, if we let the
user their email address, but required it be unique (so password reset would
work), we'd still be allowing a logged in user to set up a directory harvest
attack.

This change simplifies the system significantly, and plenty of other apps (like
YNAB) do it this way.